requirements.txt
streamlit
gensim
numpy<2.0
pdfplumber
python-docx
fuzzywuzzy
nltk
python-dateutil